opentracing-3.0: propagate span for async requests
Ensure the parent span is linked correctly by propagating the active
span to the new thread when making an asynchronous restfulWS request.

This is needed to accomodate a change in behaviour vs. JAX-RS 2.1 -
previously outgoing client filters ran on the thread that triggered the
request but now they run on a different thread.